Thesaurus compiles now, but still doesn't recognize the text. I have to debug it more.

+++ application_kit/Thesaurus/Dictionary.h
@@ -58,8 +58,8 @@ private:
 		{}
 
 		inline ~HashNode() {
-			for (Synonym *s = (Synonym *)fSynonymList->RemoveItem(0L);
-				s; s = (Synonym *)fSynonymList->RemoveItem(0L))
+			for (Synonym *s = (Synonym *)fSynonymList->RemoveItem((int32)0);
+				s; s = (Synonym *)fSynonymList->RemoveItem((int32)0))
 				delete s;
 		}
